burncast: Split podcast into smaller pieces
When actually trying out a CD burnt with burncast, I found out that
having only one huge track on the disk can be a problem if your car trip
isn't as long as a CD (i.e. <80 minutes). Scrubbing in this huge track
is not much fun, so this adds a option to cut the podcast into smaller
tracks.

To keep the time handling simple, burncast now only accepts the time in
seconds and drops support for the "00:00:00"-format.