dict.cc.py: Rework output
In my opinion, that way this script prints text was to convoluted. Also
it is not the main task for it to have pretty output, this can easily be
done by other UNIX tools.

This changes the output to be delimited by a tab character. I don't
expect that any phrases from dict.cc include a tab, so this makes the
output of dict.cc.py very friendly to parse with the usual tools. To
match the old output, run:

    $ ./dict.cc.py "spam" | column -t -s $'\t'